refactor(forks): Move helpers to forks library
