+++ src/ethereum_test_forks/__init__.py
@@ -33,8 +33,12 @@ from .helpers import (
     get_deployed_forks,
     get_development_forks,
     get_forks,
+    get_forks_with_no_descendants,
+    get_forks_with_no_parents,
     get_forks_with_solc_support,
     get_forks_without_solc_support,
+    get_from_until_fork_set,
+    get_last_descendants,
     get_transition_forks,
     transition_fork_from_to,
     transition_fork_to,
@@ -65,12 +69,16 @@ __all__ = [
     "get_transition_forks",
     "forks_from",
     "forks_from_until",
+    "get_closest_fork_with_solc_support",
     "get_deployed_forks",
     "get_development_forks",
-    "get_forks",
+    "get_forks_with_no_descendants",
+    "get_forks_with_no_parents",
     "get_forks_with_solc_support",
     "get_forks_without_solc_support",
-    "get_closest_fork_with_solc_support",
+    "get_forks",
+    "get_from_until_fork_set",
+    "get_last_descendants",
     "transition_fork_from_to",
     "transition_fork_to",
 ]
